requirements.txt
# ===========================================
# AI CHATBOT RAG - DEPENDENCIES
# ===========================================

# === Core Framework ===
fastapi>=0.115.0
uvicorn>=0.34.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0.0

# === Database (PostgreSQL) ===
sqlalchemy>=2.0.0
asyncpg>=0.30.0
psycopg2-binary>=2.9.0
alembic>=1.14.0

# === RAG Pipeline ===
langchain>=0.3.0
langchain-community>=0.3.0
langchain-google-genai>=2.0.0
langchain-chroma>=0.2.0

# === Vector Database ===
chromadb>=0.6.0

# === Embedding (Google API — không cần PyTorch) ===
google-generativeai>=0.8.0

# === PDF Processing ===
pymupdf>=1.25.0
pdfplumber>=0.11.0

# === Sparse Search (BM25) ===
rank-bm25>=0.2.0


# === Evaluation ===
ragas>=0.2.0
datasets>=3.0.0

# === Cache & Session ===
redis>=5.0.0

# === Utilities ===
numpy>=1.26.0
pandas>=2.0.0
tqdm>=4.66.0
aiofiles>=24.0.0
httpx>=0.27.0

# === Monitoring (Optional) ===
langsmith>=0.2.0
